+++ iblrig/base_choice_world.py
@@ -115,16 +115,21 @@ class ChoiceWorldSession(
     base_parameters_file = Path(__file__).parent.joinpath('base_choice_world_params.yaml')
     TrialDataModel = ChoiceWorldTrialData
 
-    def __init__(self, *args, delay_secs=0, **kwargs):
+    def __init__(self, *args, delay_mins: float = 0, **kwargs):
         super().__init__(**kwargs)
-        self.task_params['SESSION_DELAY_START'] = delay_secs
+
+        # session delay is handled in seconds internally
+        self.task_params['SESSION_DELAY_START'] = delay_mins * 60.0
+
         # init behaviour data
         self.movement_left = self.device_rotary_encoder.THRESHOLD_EVENTS[self.task_params.QUIESCENCE_THRESHOLDS[0]]
         self.movement_right = self.device_rotary_encoder.THRESHOLD_EVENTS[self.task_params.QUIESCENCE_THRESHOLDS[1]]
+
         # init counter variables
         self.trial_num = -1
         self.block_num = -1
         self.block_trial_num = -1
+
         # init the tables, there are 2 of them: a trials table and a ambient sensor data table
         self.trials_table = self.TrialDataModel.preallocate_dataframe(NTRIALS_INIT)
         self.ambient_sensor_table = pd.DataFrame(
@@ -140,12 +145,12 @@ class ChoiceWorldSession(
         """:return: argparse.parser()"""
         parser = super(ChoiceWorldSession, ChoiceWorldSession).extra_parser()
         parser.add_argument(
-            '--delay_secs',
-            dest='delay_secs',
+            '--delay_mins',
+            dest='delay_mins',
             default=0,
-            type=int,
+            type=float,
             required=False,
-            help='initial delay before starting the first trial (default: 0s)',
+            help='initial delay before starting the first trial (default: 0 min)',
         )
         parser.add_argument(
             '--remote',
@@ -264,6 +269,7 @@ class ChoiceWorldSession(
                 'bonsai_closed_loop': daction,
                 'bonsai_freeze_stim': daction,
                 'bonsai_show_center': daction,
+                'bonsai_freeze_center': daction,
             }
         )
 
@@ -434,7 +440,7 @@ class ChoiceWorldSession(
         sma.add_state(
             state_name='freeze_reward',
             state_timer=0,
-            output_actions=[self.bpod.actions.bonsai_show_center],
+            output_actions=[self.bpod.actions.bonsai_freeze_center],
             state_change_conditions={'Tup': 'reward'},
         )
         sma.add_state(
@@ -833,7 +839,7 @@ class BiasedChoiceWorldSession(ActiveChoiceWorldSession):
 
     def new_block(self):
         """
-        if block_init_5050
+        If block_init_5050
             First block has 50/50 probability of leftward stim
             is 90 trials long
         """
@@ -897,7 +903,6 @@ class TrainingChoiceWorldTrialData(ActiveChoiceWorldTrialData):
 
     training_phase: NonNegativeInt
     debias_trial: bool
-    signed_contrast: float | None = None
 
 
 class TrainingChoiceWorldSession(ActiveChoiceWorldSession):
@@ -964,25 +969,16 @@ class TrainingChoiceWorldSession(ActiveChoiceWorldSession):
         )
         return training_info['training_phase'], training_info['adaptive_reward'], training_info['adaptive_gain']
 
-    def compute_performance(self):
-        """Aggregate the trials table to compute the performance of the mouse on each contrast."""
-        self.trials_table['signed_contrast'] = self.trials_table.contrast * self.trials_table.position
-        performance = self.trials_table.groupby(['signed_contrast']).agg(
-            last_50_perf=pd.NamedAgg(column='trial_correct', aggfunc=lambda x: np.sum(x[np.maximum(-50, -x.size) :]) / 50),
-            ntrials=pd.NamedAgg(column='trial_correct', aggfunc='count'),
-        )
-        return performance
-
-    def check_training_phase(self):
+    def check_training_phase(self) -> bool:
         """Check if the mouse is ready to move to the next training phase."""
         move_on = False
         if self.training_phase == 0:  # each of the -1, -.5, .5, 1 contrast should be above 80% perf to switch
-            performance = self.compute_performance()
+            performance = choiceworld.compute_performance(self.trials_table)
             passing = performance[np.abs(performance.index) >= 0.5]['last_50_perf']
             if np.all(passing > 0.8) and passing.size == 4:
                 move_on = True
         elif self.training_phase == 1:  # each of the -.25, .25 should be above 80% perf to switch
-            performance = self.compute_performance()
+            performance = choiceworld.compute_performance(self.trials_table)
             passing = performance[np.abs(performance.index) == 0.25]['last_50_perf']
             if np.all(passing > 0.8) and passing.size == 2:
                 move_on = True
@@ -992,35 +988,47 @@ class TrainingChoiceWorldSession(ActiveChoiceWorldSession):
         if move_on:
             self.training_phase = np.minimum(5, self.training_phase + 1)
             log.warning(f'Moving on to training phase {self.training_phase}, {self.trial_num}')
+        return move_on
 
     def next_trial(self):
         # update counters
         self.trial_num += 1
         self.var['training_phase_trial_counts'][self.training_phase] += 1
+
         # check if the subject graduates to a new training phase
         self.check_training_phase()
+
         # draw the next trial
         signed_contrast = choiceworld.draw_training_contrast(self.training_phase)
         position = self.task_params.STIM_POSITIONS[int(np.sign(signed_contrast) == 1)]
         contrast = np.abs(signed_contrast)
-        # debiasing: if the previous trial was incorrect and easy repeat the trial
+
+        # debiasing: if the previous trial was incorrect, not a no-go and easy
         if self.task_params.DEBIAS and self.trial_num >= 1 and self.training_phase < 5:
             last_contrast = self.trials_table.loc[self.trial_num - 1, 'contrast']
-            do_debias_trial = (self.trials_table.loc[self.trial_num - 1, 'trial_correct'] != 1) and last_contrast >= 0.5
+            do_debias_trial = (
+                (self.trials_table.loc[self.trial_num - 1, 'trial_correct'] != 1)
+                and (self.trials_table.loc[self.trial_num - 1, 'response_side'] != 0)
+                and last_contrast >= 0.5
+            )
             self.trials_table.at[self.trial_num, 'debias_trial'] = do_debias_trial
             if do_debias_trial:
-                iresponse = np.logical_and(
-                    ~self.trials_table['response_side'].isna(), self.trials_table['response_side'] != 0
-                )  # trials that had a response
+                # indices of trials that had a response
+                iresponse = np.logical_and(self.trials_table['response_side'].notna(), self.trials_table['response_side'] != 0)
+                iresponse = iresponse.index[iresponse]
+
                 # takes the average of right responses over last 10 response trials
-                average_right = np.mean(self.trials_table['response_side'][iresponse[-np.maximum(10, iresponse.size) :]] == 1)
-                # the next probability of next stimulus being on the left is a draw from a normal distribution
-                # centered on average right with sigma 0.5. If it is less than 0.5 the next stimulus will be on the left
+                average_right = (self.trials_table['response_side'][iresponse[-np.minimum(10, iresponse.size) :]] == 1).mean()
+
+                # the probability of the next stimulus being on the left is a draw from a normal distribution centered
+                # on the average right with sigma 0.5 - if it is less than 0.5 the next stimulus will be on the left.
                 position = self.task_params.STIM_POSITIONS[int(np.random.normal(average_right, 0.5) >= 0.5)]
+
                 # contrast is the last contrast
                 contrast = last_contrast
         else:
             self.trials_table.at[self.trial_num, 'debias_trial'] = False
+
         # save and send trial info to bonsai
         self.draw_next_trial_info(pleft=self.task_params.PROBABILITY_LEFT, position=position, contrast=contrast)
         self.trials_table.at[self.trial_num, 'training_phase'] = self.training_phase