+++ iblrig/gui/wizard.py
@@ -10,6 +10,7 @@ import sys
 import traceback
 from collections import OrderedDict
 from dataclasses import dataclass
+from importlib.metadata import entry_points
 from importlib.util import module_from_spec, spec_from_file_location
 from pathlib import Path
 
@@ -132,6 +133,11 @@ class RigWizardModel:
             tasks.extend(sorted([p for p in Path(iblrig_custom_tasks.__file__).parent.rglob('task.py')]))
         self.all_tasks = OrderedDict({p.parts[-2]: p for p in tasks})
 
+        # include external tasks registered as plugins
+        for plugin in sorted(entry_points(group='iblrig.plugins'), key=lambda ep: ep.name):
+            if plugin.name.startswith('task_') and issubclass(session := plugin.load(), BaseSession):
+                self.all_tasks[session.protocol_name] = session.get_task_file()
+
         # get the subjects from iterating over folders in the the iblrig data path
         if self.iblrig_settings['iblrig_local_data_path'] is None:
             self.all_subjects = [self.test_subject_name]
@@ -248,10 +254,13 @@ class RigWizardModel:
                 )
             QtWidgets.QMessageBox().critical(None, 'Error', f'{message}\n\n{solution}')
 
-        # get subjects from Alyx: this is the set of subjects that are alive and not stock in the lab defined in settings
-        rest_subjects = self.alyx.rest(
-            'subjects', 'list', alive=True, stock=False, lab=self.iblrig_settings['ALYX_LAB'], no_cache=True
-        )
+        # get subjects from Alyx: this is the set of subjects that are alive and in the lab defined in settings
+        # stock subjects are excluded, unless the user is stock manager
+        kwargs = {'alive': True, 'lab': self.iblrig_settings['ALYX_LAB'], 'no_cache': True}
+        is_stock_manager = any(self.alyx.rest('subjects', 'list', responsible_user=self.user, stock=True, limit=1, **kwargs))
+        if not is_stock_manager:
+            kwargs['stock'] = False
+        rest_subjects = self.alyx.rest('subjects', 'list', **kwargs)
         self.all_subjects.remove(self.test_subject_name)
         self.all_subjects = [self.test_subject_name] + sorted(set(self.all_subjects + [s['nickname'] for s in rest_subjects]))
 
@@ -425,12 +434,12 @@ class RigWizard(QtWidgets.QMainWindow, Ui_wizard):
             msg_box.setWindowTitle('IBLRIG System Validation')
             msg_box.setIcon(QtWidgets.QMessageBox().Warning)
             msg_box.setTextFormat(QtCore.Qt.TextFormat.RichText)
-            text = f"The following issue{'s were' if len(results) > 1 else ' was'} detected:"
+            text = f'The following issue{"s were" if len(results) > 1 else " was"} detected:'
             for result in results:
                 text = (
-                    text + f"<br><br>\n"
-                    f"<b>{'Warning' if result.status == Status.WARN else 'Failure'}:</b> {result.message}<br>\n"
-                    f"{('<b>Suggestion:</b> ' + result.solution) if result.solution is not None else ''}"
+                    text + f'<br><br>\n'
+                    f'<b>{"Warning" if result.status == Status.WARN else "Failure"}:</b> {result.message}<br>\n'
+                    f'{("<b>Suggestion:</b> " + result.solution) if result.solution is not None else ""}'
                 )
             text = text + '<br><br>\nPlease refer to the System Validation tool for more details.'
             msg_box.setText(text)
@@ -568,10 +577,7 @@ class RigWizard(QtWidgets.QMainWindow, Ui_wizard):
         box.setModal(False)
         box.setWindowTitle('Training Level')
         box.setText(
-            f'{session_path}\n\n'
-            f'training phase:\t{training_phase}\n'
-            f'reward:\t{reward_amount:.2f} uL\n'
-            f'stimulus gain:\t{stim_gain}'
+            f'{session_path}\n\ntraining phase:\t{training_phase}\nreward:\t{reward_amount:.2f} uL\nstimulus gain:\t{stim_gain}'
         )
         if self.uiComboTask.currentText() == '_iblrig_tasks_trainingChoiceWorld':
             box.setStandardButtons(QtWidgets.QMessageBox.Apply | QtWidgets.QMessageBox.Close)
@@ -862,9 +868,9 @@ class RigWizard(QtWidgets.QMainWindow, Ui_wizard):
                     widget.setMinimum(0)
                     widget.setMaximum(11)
 
-                case 'delay_secs':
-                    label = 'Initial Delay, s'
-                    widget.setMaximum(86400)
+                case 'delay_mins':
+                    label = 'Initial Delay, min'
+                    widget.setMaximum(60)
 
                 case 'training_phase':
                     widget.setSpecialValueText('automatic')
@@ -1146,9 +1152,9 @@ class RigWizard(QtWidgets.QMainWindow, Ui_wizard):
                 answer = QtWidgets.QMessageBox.question(
                     self,
                     'Is this a dud?',
-                    f"The session consisted of only {ntrials:d} trial"
-                    f"{'s' if ntrials > 1 else ''} and appears to be a dud.\n\n"
-                    f"Should it be deleted?",
+                    f'The session consisted of only {ntrials:d} trial'
+                    f'{"s" if ntrials > 1 else ""} and appears to be a dud.\n\n'
+                    f'Should it be deleted?',
                 )
                 if answer == QtWidgets.QMessageBox.Yes:
                     shutil.rmtree(self.model.session_folder)